ceph: update helm version
Currently, Helm v2.5.1 is used, but this version is released in 2017
and a little outdated.
To use the latest version of helm v2.x.x would be beneficial
as we can use more useful Helm commands like `helm template`.

This commit updates Helm version and values.yaml
because helm v2.16.7 does not allow to use undefined values in values.yaml
when using `helm lint --strict` command.

Signed-off-by: binoue <banji-inoue@cybozu.co.jp>